thử vòng lặp và load sprite

+++ src/uet/oop/bomberman/entities/characters/Bomber.java
@@ -9,7 +9,7 @@ import uet.oop.bomberman.entities.Entity;
 
 public class Bomber extends Entity {
 
-    public Bomber(int x, int y, Image img) {
+    public Bomber(double x, double y, Image img) {
         super( x, y, img);
     }
